requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
redis==5.0.1
chromadb==0.4.18
discord.py==2.3.2
aiohttp==3.9.1
python-multipart==0.0.6
jinja2==3.1.2
supabase==2.0.3
